+++ scrapy/core/downloader/contextfactory.py
@@ -79,13 +79,6 @@ class _ScrapyClientContextFactory(BrowserLikePolicyForHTTPS):
             self.tls_ciphers = AcceptableCiphers.fromOpenSSLCipherString(tls_ciphers)
         else:
             self.tls_ciphers = DEFAULT_CIPHERS
-        with _filter_method_warning():
-            self._certificate_options = CertificateOptions(
-                method=self._ssl_method,
-                fixBrokenPeers=True,
-                acceptableCiphers=self.tls_ciphers,
-            )
-        self._ctx = self._get_context()
         self._verify_certificates = verify_certificates
 
     @classmethod
@@ -110,31 +103,46 @@ class _ScrapyClientContextFactory(BrowserLikePolicyForHTTPS):
             **kwargs,
         )
 
+    # should be removed together with ScrapyClientContextFactory
     def getCertificateOptions(self) -> CertificateOptions:  # pragma: no cover
-        return self._certificate_options
+        return self._get_cert_options()
 
-    # kept for old-style HTTP/1.0 downloader context twisted calls,
-    # e.g. connectSSL()
-    def getContext(self, hostname: Any = None, port: Any = None) -> SSL.Context:
-        return self._ctx
+    def _get_cert_options(self) -> CertificateOptions:
+        with _filter_method_warning():
+            return CertificateOptions(  # type: ignore[no-any-return]
+                method=self._ssl_method,
+                fixBrokenPeers=True,
+                acceptableCiphers=self.tls_ciphers,
+            )
+
+    # should be removed together with ScrapyClientContextFactory
+    def getContext(
+        self, hostname: Any = None, port: Any = None
+    ) -> SSL.Context:  # pragma: no cover
+        return self._get_context()
 
     def _get_context(self) -> SSL.Context:
-        ctx = self._certificate_options.getContext()
+        cert_options = self._get_cert_options()
+        ctx: SSL.Context = cert_options.getContext()
         ctx.set_options(0x4)  # OP_LEGACY_SERVER_CONNECT
         return ctx
 
     def creatorForNetloc(self, hostname: bytes, port: int) -> ClientTLSOptions:
         if not self._verify_certificates:
+            # Our options class is needed to skip verification errors
             if TWISTED_TLS_NEW_IMPL:
                 return _ScrapyClientTLSOptions26(
-                    self._certificate_options._makeTLSConnection,
+                    self._get_cert_options()._makeTLSConnection,
                     hostname.decode("ascii"),
                     verbose_logging=self.tls_verbose_logging,
                 )
-            return _ScrapyClientTLSOptions(hostname.decode("ascii"), self._ctx)  # type: ignore[no-untyped-call]
-        # Note that this doesn't use self._ctx
+            return _ScrapyClientTLSOptions(
+                hostname.decode("ascii"), self._get_context()
+            )  # type: ignore[no-untyped-call]
+        # Otherwise use the normal Twisted function.
+        # Note that this doesn't use self._get_context().
         with _filter_method_warning():
-            return optionsForClientTLS(
+            return optionsForClientTLS(  # type: ignore[no-any-return]
                 hostname=hostname.decode("ascii"),
                 extraCertificateOptions={
                     "method": self._ssl_method,
@@ -185,7 +193,7 @@ class BrowserLikeContextFactory(_ScrapyClientContextFactory):
 
     def creatorForNetloc(self, hostname: bytes, port: int) -> ClientTLSOptions:
         with _filter_method_warning():
-            return optionsForClientTLS(
+            return optionsForClientTLS(  # type: ignore[no-any-return]
                 hostname=hostname.decode("ascii"),
                 extraCertificateOptions={"method": self._ssl_method},
             )