+++ tests/test_downloader_handler_twisted_http2.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import json
 from typing import TYPE_CHECKING, Any
 
 import pytest
@@ -11,18 +10,23 @@ from twisted.web.http import H2_ENABLED
 
 from scrapy import Spider
 from scrapy.crawler import Crawler
-from scrapy.exceptions import NotConfigured, UnsupportedURLSchemeError
+from scrapy.exceptions import (
+    DownloadFailedError,
+    NotConfigured,
+    UnsupportedURLSchemeError,
+)
 from scrapy.http import Request
-from scrapy.utils.defer import deferred_f_from_coro_f, maybe_deferred_to_future
+from scrapy.utils.defer import maybe_deferred_to_future
 from tests.test_downloader_handlers_http_base import (
     TestHttpProxyBase,
-    TestHttps11Base,
+    TestHttpsBase,
     TestHttpsCustomCiphersBase,
     TestHttpsInvalidDNSIdBase,
     TestHttpsInvalidDNSPatternBase,
     TestHttpsWrongHostnameBase,
     TestHttpWithCrawlerBase,
 )
+from tests.utils.decorators import coroutine_test
 
 if TYPE_CHECKING:
     from scrapy.core.downloader.handlers import DownloadHandlerProtocol
@@ -57,29 +61,16 @@ def test_not_configured_without_reactor() -> None:
         H2DownloadHandler.from_crawler(crawler)
 
 
-class TestHttps2(H2DownloadHandlerMixin, TestHttps11Base):
-    HTTP2_DATALOSS_SKIP_REASON = "Content-Length mismatch raises InvalidBodyLengthError"
+class TestHttp2(H2DownloadHandlerMixin, TestHttpsBase):
+    http2 = True
+    handler_supports_http2_dataloss = False
 
-    @deferred_f_from_coro_f
+    @coroutine_test
     async def test_protocol(self, mockserver: MockServer) -> None:
-        request = Request(
-            mockserver.url("/host", is_secure=self.is_secure), method="GET"
-        )
+        request = Request(mockserver.url("/host", is_secure=self.is_secure))
         async with self.get_dh() as download_handler:
             response = await download_handler.download_request(request)
         assert response.protocol == "h2"
-
-    def test_download_cause_data_loss(self) -> None:  # type: ignore[override]
-        pytest.skip(self.HTTP2_DATALOSS_SKIP_REASON)
-
-    def test_download_cause_data_loss_double_warning(self) -> None:  # type: ignore[override]
-        pytest.skip(self.HTTP2_DATALOSS_SKIP_REASON)
-
-    def test_download_allow_data_loss(self) -> None:  # type: ignore[override]
-        pytest.skip(self.HTTP2_DATALOSS_SKIP_REASON)
-
-    def test_download_allow_data_loss_via_setting(self) -> None:  # type: ignore[override]
-        pytest.skip(self.HTTP2_DATALOSS_SKIP_REASON)
 
     def test_download_conn_failed(self) -> None:  # type: ignore[override]
         # Unlike HTTP11DownloadHandler which raises it from download_request()
@@ -89,12 +80,6 @@ class TestHttps2(H2DownloadHandlerMixin, TestHttps11Base):
         # DOWNLOAD_TIMEOUT.
         pytest.skip("The handler doesn't properly reraise ConnectionRefusedError")
 
-    def test_download_conn_lost(self) -> None:  # type: ignore[override]
-        pytest.skip(self.HTTP2_DATALOSS_SKIP_REASON)
-
-    def test_download_conn_aborted(self) -> None:  # type: ignore[override]
-        pytest.skip(self.HTTP2_DATALOSS_SKIP_REASON)
-
     def test_download_dns_error(self) -> None:  # type: ignore[override]
         # Unlike HTTP11DownloadHandler which raises it from download_request()
         # (without any special handling), here DNSLookupError (raised in
@@ -103,7 +88,7 @@ class TestHttps2(H2DownloadHandlerMixin, TestHttps11Base):
         # DOWNLOAD_TIMEOUT.
         pytest.skip("The handler doesn't properly reraise DNSLookupError")
 
-    @deferred_f_from_coro_f
+    @coroutine_test
     async def test_concurrent_requests_same_domain(
         self, mockserver: MockServer
     ) -> None:
@@ -118,7 +103,7 @@ class TestHttps2(H2DownloadHandlerMixin, TestHttps11Base):
             assert response2.headers["Content-Length"] == b"79"
 
     @pytest.mark.xfail(reason="https://github.com/python-hyper/h2/issues/1247")
-    @deferred_f_from_coro_f
+    @coroutine_test
     async def test_connect_request(self, mockserver: MockServer) -> None:
         request = Request(
             mockserver.url("/file", is_secure=self.is_secure), method="CONNECT"
@@ -127,7 +112,7 @@ class TestHttps2(H2DownloadHandlerMixin, TestHttps11Base):
             response = await download_handler.download_request(request)
         assert response.body == b""
 
-    @deferred_f_from_coro_f
+    @coroutine_test
     async def test_custom_content_length_good(self, mockserver: MockServer) -> None:
         request = Request(mockserver.url("/contentlength", is_secure=self.is_secure))
         custom_content_length = str(len(request.body))
@@ -136,7 +121,7 @@ class TestHttps2(H2DownloadHandlerMixin, TestHttps11Base):
             response = await download_handler.download_request(request)
         assert response.text == custom_content_length
 
-    @deferred_f_from_coro_f
+    @coroutine_test
     async def test_custom_content_length_bad(self, mockserver: MockServer) -> None:
         request = Request(mockserver.url("/contentlength", is_secure=self.is_secure))
         actual_content_length = str(len(request.body))
@@ -156,38 +141,33 @@ class TestHttps2(H2DownloadHandlerMixin, TestHttps11Base):
             )
         )
 
-    @deferred_f_from_coro_f
-    async def test_duplicate_header(self, mockserver: MockServer) -> None:
-        request = Request(mockserver.url("/echo", is_secure=self.is_secure))
-        header, value1, value2 = "Custom-Header", "foo", "bar"
-        request.headers.appendlist(header, value1)
-        request.headers.appendlist(header, value2)
+    @coroutine_test
+    async def test_data_loss_handling(self, mockserver: MockServer) -> None:
+        request = Request(mockserver.url("/broken", is_secure=self.is_secure))
         async with self.get_dh() as download_handler:
-            response = await download_handler.download_request(request)
-        assert json.loads(response.text)["headers"][header] == [value1, value2]
+            with pytest.raises(DownloadFailedError):
+                await download_handler.download_request(request)
 
 
-class TestHttps2WrongHostname(H2DownloadHandlerMixin, TestHttpsWrongHostnameBase):
+class TestHttp2WrongHostname(H2DownloadHandlerMixin, TestHttpsWrongHostnameBase):
     pass
 
 
-class TestHttps2InvalidDNSId(H2DownloadHandlerMixin, TestHttpsInvalidDNSIdBase):
+class TestHttp2InvalidDNSId(H2DownloadHandlerMixin, TestHttpsInvalidDNSIdBase):
     pass
 
 
-class TestHttps2InvalidDNSPattern(
+class TestHttp2InvalidDNSPattern(
     H2DownloadHandlerMixin, TestHttpsInvalidDNSPatternBase
 ):
     pass
 
 
-class TestHttps2CustomCiphers(H2DownloadHandlerMixin, TestHttpsCustomCiphersBase):
+class TestHttp2CustomCiphers(H2DownloadHandlerMixin, TestHttpsCustomCiphersBase):
     pass
 
 
 class TestHttp2WithCrawler(TestHttpWithCrawlerBase):
-    """HTTP 2.0 test case with MockServer"""
-
     @property
     def settings_dict(self) -> dict[str, Any] | None:
         return {
@@ -212,11 +192,11 @@ class TestHttp2WithCrawler(TestHttpWithCrawlerBase):
         pytest.skip("headers_received support is not implemented")
 
 
-class TestHttps2Proxy(H2DownloadHandlerMixin, TestHttpProxyBase):
+class TestHttp2Proxy(H2DownloadHandlerMixin, TestHttpProxyBase):
     is_secure = True
     expected_http_proxy_request_body = b"/"
 
-    @deferred_f_from_coro_f
+    @coroutine_test
     async def test_download_with_proxy_https_timeout(
         self, proxy_mockserver: ProxyEchoMockServer
     ) -> None:
@@ -225,7 +205,7 @@ class TestHttps2Proxy(H2DownloadHandlerMixin, TestHttpProxyBase):
                 super().test_download_with_proxy_https_timeout(proxy_mockserver)  # type: ignore[arg-type]
             )
 
-    @deferred_f_from_coro_f
+    @coroutine_test
     async def test_download_with_proxy_without_http_scheme(
         self, proxy_mockserver: ProxyEchoMockServer
     ) -> None: